+++ backend/canisters/notifications_index/impl/src/updates/c2c_user_index.rs
@@ -34,6 +34,17 @@ fn c2c_user_index_impl(args: Args, state: &mut RuntimeState) -> Response {
                         *now,
                     );
                 }
+                UserIndexEvent::BotEndpointUpdated(user_id, endpoint) => {
+                    state.data.bot_endpoints.insert(user_id, endpoint.clone());
+                    state.push_event_to_notifications_canisters(
+                        NotificationsIndexEvent::BotEndpointUpdated(user_id, endpoint.clone()),
+                        *now,
+                    );
+                }
+                UserIndexEvent::BotRemoved(user_id) => {
+                    state.data.bot_endpoints.remove(&user_id);
+                    state.push_event_to_notifications_canisters(NotificationsIndexEvent::BotRemoved(user_id), *now);
+                }
             }
         }
     }